requirements.txt.txt
inference-gpu
supervision
numpy
opencv-python